+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 1.0.0
 
 - Create select_bottom_list package.
+- Example for select_bottom_list use package.